pyproject.toml
[project]
name = "Medical NLP Summarization"
version = "0.1.0"
description = ""
authors = [
    {name = "Rashi Mishra",email = "rashimishra193@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.66.2,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "flask (>=3.1.0,<4.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"